django/requirements.txt
Django==4.2
psycopg2-binary==2.9.6
django-environ==0.8.1
djangorestframework
